+++ scripts/tests/single_artifact_multiple_policies.py
@@ -19,7 +19,6 @@ from resourceapi import ResourceApi
 from idsapi import IdsApi
 import requests
 import pprint
-import json
 import sys